requirements.txt
asgiref==3.2.3
Django==3.0.1
django-annoying==0.10.6
django-appconf==1.0.3
django-imagekit==4.0.2
pilkit==2.0
Pillow==7.0.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
